requirements.txt
google-api-python-client>=2.70.0
pandas>=1.5.0
google-auth>=2.15.0
